[SYCL][NFC] Promote new-style property engine base to sycl::detail
