+++ sycl/test/abi/sycl_classes_abi_neutral_test.cpp
@@ -16,7 +16,7 @@
 // member is not crossing ABI boundary. All current exclusions are listed below.
 
 // CHECK: 0 | struct sycl::ext::oneapi::experimental::build_options
-// CHECK-NEXT:         0 |   struct sycl::ext::oneapi::experimental::detail::run_time_property_key
+// CHECK-NEXT:         0 |   struct sycl::detail::run_time_property_key
 // CHECK:              0 |   class std::vector<class std::basic_string<char> > opts
 // CHECK-NEXT:         0 |     struct std::_Vector_base<class std::basic_string<char>, class std::allocator<class std::basic_string<char> > > (base)
 // CHECK-NEXT:         0 |       struct std::_Vector_base<class std::basic_string<char>, class std::allocator<class std::basic_string<char> > >::_Vector_impl _M_impl
@@ -25,7 +25,7 @@
 // CHECK-NEXT:         0 |         {{(struct std::_Vector_base<class std::basic_string<char>, class std::allocator<class std::basic_string<char> > >::_Vector_impl_data \(base\)|pointer _M_start)}}
 
 // CHECK: 0 | struct sycl::ext::oneapi::experimental::include_files
-// CHECK-NEXT:         0 |   struct sycl::ext::oneapi::experimental::detail::run_time_property_key
+// CHECK-NEXT:         0 |   struct sycl::detail::run_time_property_key
 // CHECK:              0 |   class std::vector<struct std::pair<class std::basic_string<char>, class std::basic_string<char> > > record
 // CHECK-NEXT:         0 |     struct std::_Vector_base<struct std::pair<class std::basic_string<char>, class std::basic_string<char> >, class std::allocator<struct std::pair<class std::basic_string<char>, class std::basic_string<char> > > > (base)
 // CHECK-NEXT:         0 |       struct std::_Vector_base<struct std::pair<class std::basic_string<char>, class std::basic_string<char> >, class std::allocator<struct std::pair<class std::basic_string<char>, class std::basic_string<char> > > >::_Vector_impl _M_impl
@@ -34,7 +34,7 @@
 // CHECK-NEXT:         0 |         {{(struct std::_Vector_base<struct std::pair<class std::basic_string<char>, class std::basic_string<char> >, class std::allocator<struct std::pair<class std::basic_string<char>, class std::basic_string<char> > > >::_Vector_impl_data \(base\)|pointer _M_start)}}
 
 // CHECK: 0 | struct sycl::ext::oneapi::experimental::registered_names
-// CHECK-NEXT:         0 |   struct sycl::ext::oneapi::experimental::detail::run_time_property_key
+// CHECK-NEXT:         0 |   struct sycl::detail::run_time_property_key
 // CHECK:              0 |   class std::vector<class std::basic_string<char> > names
 // CHECK-NEXT:         0 |     struct std::_Vector_base<class std::basic_string<char>, class std::allocator<class std::basic_string<char> > > (base)
 // CHECK-NEXT:         0 |       struct std::_Vector_base<class std::basic_string<char>, class std::allocator<class std::basic_string<char> > >::_Vector_impl _M_impl
